requirements.txt
streamlit>=1.36
notion-client>=3.0
PyYAML>=6.0
Pillow>=10.0
